requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
requests>=2.31.0
yt-dlp>=2024.0.0
aiocryptopay>=0.4.0
